pyproject.toml
[project]
name = "crypto-intel-bot"
version = "1.0.0"
description = "Unified Crypto Intelligence Telegram Bot — market analysis, trading signals, social intelligence, and liquidation alerts"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.10",
    "anthropic>=0.52.0",
    "aiohttp>=3.9",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "mplfinance>=0.12",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
